requirements.txt
# requirements.txt
flask==2.0.1
werkzeug==2.0.2
python-dotenv==1.0.0
requests==2.28.2